Adds lastModified property and prevents caching if file already exists (Android)

+++ file_picker/android/src/main/java/com/mr/flutter/plugin/filepicker/FileUtils.java
@@ -17,8 +17,11 @@ import android.webkit.MimeTypeMap;
 
 import androidx.annotation.Nullable;
 
+import java.io.BufferedInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -31,106 +34,6 @@ public class FileUtils {
 
     private static final String TAG = "FilePickerUtils";
     private static final String PRIMARY_VOLUME_NAME = "primary";
-
-    public static String getPath(final Uri uri, final Context context) {
-        final boolean isKitKat = Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT;
-        if (isKitKat) {
-            return getForApi19(context, uri);
-        } else if ("content".equalsIgnoreCase(uri.getScheme())) {
-            if (isGooglePhotosUri(uri)) {
-                return uri.getLastPathSegment();
-            }
-            return getDataColumn(context, uri, null, null);
-        } else if ("file".equalsIgnoreCase(uri.getScheme())) {
-            return uri.getPath();
-        }
-        return null;
-    }
-
-    @TargetApi(19)
-    @SuppressWarnings("deprecation")
-    private static String getForApi19(final Context context, final Uri uri) {
-        Log.e(TAG, "Getting for API 19 or above" + uri);
-        if (DocumentsContract.isDocumentUri(context, uri)) {
-            Log.e(TAG, "Document URI");
-            if (isExternalStorageDocument(uri)) {
-                Log.e(TAG, "External Document URI");
-                final String docId = DocumentsContract.getDocumentId(uri);
-                final String[] split = docId.split(":");
-                final String type = split[0];
-                if ("primary".equalsIgnoreCase(type)) {
-                    Log.e(TAG, "Primary External Document URI");
-                    return Environment.getExternalStorageDirectory() + (split.length > 1 ? ("/" + split[1]) : "");
-                }
-            } else if (isDownloadsDocument(uri)) {
-                Log.e(TAG, "Downloads External Document URI");
-                String id = DocumentsContract.getDocumentId(uri);
-
-                if (!TextUtils.isEmpty(id)) {
-                    if (id.startsWith("raw:")) {
-                        return id.replaceFirst("raw:", "");
-                    }
-                    final String[] contentUriPrefixesToTry = new String[]{
-                            "content://downloads/public_downloads",
-                            "content://downloads/my_downloads",
-                            "content://downloads/all_downloads"
-                    };
-                    if (id.contains(":")) {
-                        id = id.split(":")[1];
-                    }
-                    for (final String contentUriPrefix : contentUriPrefixesToTry) {
-                        try {
-                            final Uri contentUri = ContentUris.withAppendedId(Uri.parse(contentUriPrefix), Long.valueOf(id));
-                            final String path = getDataColumn(context, contentUri, null, null);
-                            if (path != null) {
-                                return path;
-                            }
-                        } catch (final Exception e) {
-                            Log.e(TAG, "Something went wrong while retrieving document path: " + e.toString());
-                            return null;
-                        }
-                    }
-
-                }
-            } else if (isMediaDocument(uri)) {
-                Log.e(TAG, "Media Document URI");
-                final String docId = DocumentsContract.getDocumentId(uri);
-                final String[] split = docId.split(":");
-                final String type = split[0];
-
-                Uri contentUri = null;
-                if ("image".equals(type)) {
-                    Log.i(TAG, "Image Media Document URI");
-                    contentUri = MediaStore.Images.Media.EXTERNAL_CONTENT_URI;
-                } else if ("video".equals(type)) {
-                    Log.i(TAG, "Video Media Document URI");
-                    contentUri = MediaStore.Video.Media.EXTERNAL_CONTENT_URI;
-                } else if ("audio".equals(type)) {
-                    Log.i(TAG, "Audio Media Document URI");
-                    contentUri = MediaStore.Audio.Media.EXTERNAL_CONTENT_URI;
-                }
-
-                final String selection = MediaStore.Images.Media._ID + "=?";
-                final String[] selectionArgs = new String[]{
-                        split[1]
-                };
-
-                return getDataColumn(context, contentUri, selection, selectionArgs);
-            }
-        } else if ("content".equalsIgnoreCase(uri.getScheme())) {
-            Log.e(TAG, "NO DOCUMENT URI - CONTENT: " + uri.getPath());
-            if (isGooglePhotosUri(uri)) {
-                return uri.getLastPathSegment();
-            } else if (isDropBoxUri(uri)) {
-                return null;
-            }
-            return getDataColumn(context, uri, null, null);
-        } else if ("file".equalsIgnoreCase(uri.getScheme())) {
-            Log.e(TAG, "No DOCUMENT URI - FILE: " + uri.getPath());
-            return uri.getPath();
-        }
-        return null;
-    }
 
     public static String[] getMimeTypes(final ArrayList<String> allowedExtensions) {
 
@@ -151,29 +54,6 @@ public class FileUtils {
         }
         Log.d(TAG, "Allowed file extensions mimes: " + mimes);
         return mimes.toArray(new String[0]);
-    }
-
-    private static String getDataColumn(final Context context, final Uri uri, final String selection,
-                                        final String[] selectionArgs) {
-        Cursor cursor = null;
-        final String column = MediaStore.Images.Media.DATA;
-        final String[] projection = {
-                column
-        };
-        try {
-            cursor = context.getContentResolver().query(uri, projection, selection, selectionArgs,
-                    null);
-            if (cursor != null && cursor.moveToFirst()) {
-                final int index = cursor.getColumnIndexOrThrow(column);
-                return cursor.getString(index);
-            }
-        } catch (final Exception ex) {
-        } finally {
-            if (cursor != null) {
-                cursor.close();
-            }
-        }
-        return null;
     }
 
     public static String getFileName(Uri uri, final Context context) {
@@ -245,41 +125,59 @@ public class FileUtils {
         final String path = context.getCacheDir().getAbsolutePath() + "/file_picker/" + (fileName != null ? fileName : new Random().nextInt(100000));
 
         final File file = new File(path);
-        file.getParentFile().mkdirs();
 
-        try {
-            fos = new FileOutputStream(path);
+        if(file.exists()) {
+            int size = (int) file.length();
+            byte[] bytes = new byte[size];
+
             try {
-                final ByteArrayOutputStream out = new ByteArrayOutputStream();
-                final InputStream in = context.getContentResolver().openInputStream(uri);
-
-                final byte[] buffer = new byte[8192];
-                int len = 0;
-
-                while ((len = in.read(buffer)) >= 0) {
-                    out.write(buffer, 0, len);
-                }
-
-                fileInfo.withData(out.toByteArray());
-                out.writeTo(fos);
-                out.flush();
-            } finally {
-                fos.getFD().sync();
-            }
-        } catch (final Exception e) {
-            try {
-                fos.close();
-            } catch (final IOException | NullPointerException ex) {
+                BufferedInputStream buf = new BufferedInputStream(new FileInputStream(file));
+                buf.read(bytes, 0, bytes.length);
+                buf.close();
+            } catch (FileNotFoundException e) {
+                Log.e(TAG, "File not found: " + e.getMessage(), null);
+            } catch (IOException e) {
                 Log.e(TAG, "Failed to close file streams: " + e.getMessage(), null);
+            }
+            fileInfo.withData(bytes);
+        } else {
+
+            file.getParentFile().mkdirs();
+            try {
+                fos = new FileOutputStream(path);
+                try {
+                    final ByteArrayOutputStream out = new ByteArrayOutputStream();
+                    final InputStream in = context.getContentResolver().openInputStream(uri);
+
+                    final byte[] buffer = new byte[8192];
+                    int len = 0;
+
+                    while ((len = in.read(buffer)) >= 0) {
+                        out.write(buffer, 0, len);
+                    }
+
+                    fileInfo.withData(out.toByteArray());
+                    out.writeTo(fos);
+                    out.flush();
+                } finally {
+                    fos.getFD().sync();
+                }
+            } catch (final Exception e) {
+                try {
+                    fos.close();
+                } catch (final IOException | NullPointerException ex) {
+                    Log.e(TAG, "Failed to close file streams: " + e.getMessage(), null);
+                    return null;
+                }
+                Log.e(TAG, "Failed to retrieve path: " + e.getMessage(), null);
                 return null;
             }
-            Log.e(TAG, "Failed to retrieve path: " + e.getMessage(), null);
-            return null;
         }
 
         Log.d(TAG, "File loaded and cached at:" + path);
 
         fileInfo
+                .lastModifiedAt(file.lastModified())
                 .withPath(path)
                 .withName(fileName)
                 .withSize(Integer.parseInt(String.valueOf(file.length()/1024)))
@@ -289,7 +187,7 @@ public class FileUtils {
     }
 
     @Nullable
-    public static FileInfo getFullPathFromTreeUri(@Nullable final Uri treeUri, Context con) {
+    public static String getFullPathFromTreeUri(@Nullable final Uri treeUri, Context con) {
         if (treeUri == null) {
             return null;
         }
@@ -300,7 +198,7 @@ public class FileUtils {
         fileInfo.withUri(treeUri);
 
         if (volumePath == null) {
-            return fileInfo.withDirectory(File.separator).build();
+            return File.separator;
         }
 
         if (volumePath.endsWith(File.separator))
@@ -313,13 +211,13 @@ public class FileUtils {
 
         if (documentPath.length() > 0) {
             if (documentPath.startsWith(File.separator)) {
-                return fileInfo.withDirectory(volumePath + documentPath).build();
+                return volumePath + documentPath;
             }
             else {
-                return fileInfo.withDirectory(volumePath + File.separator + documentPath).build();
+                return volumePath + File.separator + documentPath;
             }
         } else {
-            return fileInfo.withDirectory(volumePath).build();
+            return volumePath;
         }
     }
 
@@ -373,26 +271,6 @@ public class FileUtils {
         final String[] split = docId.split(":");
         if ((split.length >= 2) && (split[1] != null)) return split[1];
         else return File.separator;
-    }
-
-    private static boolean isDropBoxUri(final Uri uri) {
-        return "com.dropbox.android.FileCache".equals(uri.getAuthority());
-    }
-
-    private static boolean isExternalStorageDocument(final Uri uri) {
-        return "com.android.externalstorage.documents".equals(uri.getAuthority());
-    }
-
-    private static boolean isDownloadsDocument(final Uri uri) {
-        return "com.android.providers.downloads.documents".equals(uri.getAuthority());
-    }
-
-    private static boolean isMediaDocument(final Uri uri) {
-        return "com.android.providers.media.documents".equals(uri.getAuthority());
-    }
-
-    private static boolean isGooglePhotosUri(final Uri uri) {
-        return "com.google.android.apps.photos.content".equals(uri.getAuthority());
     }
 
 }
